Switch back to Hash from byte[], use multiple threads for hashing.
Also some WIP on integration with the merkle tree. Some things are wrong.

+++ hedera-node/src/main/java/com/hedera/services/state/merkle/virtual/persistence/mmap/MemMapDataStore.java
@@ -410,7 +410,7 @@ public final class MemMapDataStore {
             for (int i = (size-1); i >= 0; i--) {
                 file.seek(slotOffset);
                 // read size from file
-                short size = file.readByte();
+                long size = file.readLong();
                 if (size != EMPTY) {
                     // we found some real data
                     foundFirstData = true;
